+++ docs/app/components/content/ComponentExample.vue
@@ -90,6 +90,7 @@ const wrapperContainer = ref<HTMLElement | null>(null)
 const componentContainer = ref<HTMLElement | null>(null)
 
 const { $prettier } = useNuxtApp()
+const { framework } = useFrameworks()
 const { width } = useElementSize(el)
 const config = useRuntimeConfig()
 
@@ -111,8 +112,10 @@ const code = computed(() => {
 `
   }
 
+  const source = framework.value === 'vue' && props.lang === 'vue' ? addVueImports(data.value?.code ?? '') : (data.value?.code ?? '')
+
   code += `\`\`\`${props.lang} ${props.preview ? '' : ` [${props.filename ?? data.value?.pascalName}.${props.lang}]`}${props.highlights?.length ? `{${props.highlights.join('-')}}` : ''}
-${data.value?.code ?? ''}
+${source}
 \`\`\``
 
   if (props.collapse) {