+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * docs(editor): reorder drag handle as last child in examples
 * docs(content): update filenames to be consistent
 * docs(input): fix duplicated calling code in phone number example
+* docs: add Vue imports to code examples in Vue mode
 * 
 * chore(agents): add guide for coding agents
 * chore(deps): update dependency reka-ui to v2.9.0